# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ScriptForge API
  version: 1.0.0
  description: ScriptForge 前台 / 后台 REST API 契约
paths:
...
  /api/creation/submit/:
    post:
      operationId: creation_submit_create
      description: |-
        提交创作任务

        POST /api/creation/submit/
        Body:
          {
            "theme": "family-revenge",
            "core_idea": "一句话创意描述",
            "episode_count": 30,
            "format_variant": "B",
            "audience": "",
            "reference_work": ""
          }
        Response:
          {
            "code": 0,
            "message": "success",
            "data": { "project_id": "...", "estimated_minutes": 12 }
          }
      tags:
      - creation
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT